+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-05-17
+
+### Added
+
+- Add RemainingValueTuner to transform discrete trend labels into continuous values expressing trend potential.
+- Add smoothing utilities.
+- Add label tuner example notebook.
+
+### Changed
+
+- Update documentation to include new features.
+- Update README.md to include new features.
+
+
 ## [0.2.1] - 2025-03-30
 
 ### Added